Sua lai code va bo sung sweetalert

+++ src/components/cart-item/index.js
@@ -1,19 +1,39 @@
 import React, { Component } from "react";
 import {connect} from 'react-redux';
 import { actDeleteCartItem, actUpdateQuantity } from "../../redux/actions/cart";
+import Swal from 'sweetalert2'
 
 class CartItem extends Component {
 
 
-    handleUpdateQuantity = (product,quantity) => {
+    handleUpdateQuantity = (product,flag) => {
        
-            this.props.onUpdateQuantity(product,quantity)
+            this.props.onUpdateQuantity(product,flag)
        
     }
 
     handleOnDelete = (id) => {
-        this.props.onDeleteCartItem(this.props.cartItem.id);
-        console.log(this.props.cartItem.id);
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        type: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      }).then((result) => {
+        if (result.value) {
+          this.props.onDeleteCartItem(this.props.cartItem.id);
+          console.log(this.props.cartItem.id);
+          Swal.fire(
+            'Deleted!',
+            'Your file has been deleted.',
+            'success'
+          )
+        }
+      })
+    
+       
     }
   render() {
 
@@ -29,16 +49,17 @@ class CartItem extends Component {
           />
         </td>
         <td style={{ fontSize: 25 }}>{name}</td>
-        <td>{this.props.cartItem.price.toLocaleString()}</td>
+        <td> $ {this.props.cartItem.price.toLocaleString()}</td>
         <td>
-          {this.props.quantity}
+          <span className="badge badge-success">{this.props.cartItem.quantity}</span>
+          
           <div className="btn-group">
-            <button className="btn btn-info border-right" onClick={() => this.handleUpdateQuantity(this.props.cartItem,false)}>-</button>
-            {this.props.cartItem.quantity}
+            <button className="btn btn-info border-right " onClick={() => this.handleUpdateQuantity(this.props.cartItem,false)}>-</button>
+           
             <button className="btn btn-info border-left" onClick={() => this.handleUpdateQuantity(this.props.cartItem,true)}>+</button>
           </div>
         </td>
-        <td>{(this.props.cartItem.price * this.props.cartItem.quantity).toLocaleString()}</td>
+        <td>$ {(this.props.cartItem.price * this.props.cartItem.quantity).toLocaleString()}</td>
         <td>
           <button className="btn btn-info" onClick={this.handleOnDelete}>x</button>
         </td>
